complete conversion to new accessibility-compliant UI with daisyUI
