--- chrome/popup.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <link href="css/tailwind.min.css" rel="stylesheet">
-</head>
-
-<body style="padding: 1em;">
-    <h3 id="scionsupport" class="text-lg text-left font-bold mb-3 mt-3"></h3>
-    <div id="domainlist" class="w-100 mb-4"></div>
-    <hr />
-
-    <div class="flex justify-center mb-2" id="scionModePreference">
-        <label for="toogleA" class="flex items-center cursor-pointer">
-            <div id="maindomain" class="mr-3 text-gray-700 font-medium text-sm ">
-
-            </div>
-            <div class="flex flex-col" style="width: 95px;" class="items-center">
-                <!-- toggle -->
-                <div class="relative" id="togglePerSiteStrictModeContainer" style="margin: auto;">
-                    <!-- input -->
-                    <input id="togglePerSiteStrictMode" type="checkbox" class="sr-only"/>
-                    <!-- line -->
-                    <div id="lineRunning" class="w-8 h-5 bg-gray-400 rounded-full shadow-inner"></div>
-                    <!-- dot -->
-                    <div class="dot absolute w-5 h-5 bg-white rounded-full shadow -left-1 -top-0 transition"></div>
-                </div>
-                <span id="scionmode" style="font-size: 11px; text-align: center;"></span>
-            </div>
-            <!-- label -->
-        </label>
-    </div>
-    <hr />
-
-    <div class="flex">
-        <button
-            class="px-2 py-1 rounded-md text-sm font-medium border-0 focus:outline-none focus:ring transition text-blue-600 hover:text-blue-800 active:bg-blue-200 focus:ring-blue-300"
-            id="button-options">Manage Preferences...</button>
-    </div>
-    <hr />
-
-    <div class="ac">
-        <input class="ac-input" id="ac-1" name="ac-1" type="checkbox" />
-        <label id="ac-label-main"
-            class="ac-label text-sm font-medium border-0 focus:outline-none focus:ring transition text-blue-600 hover:text-blue-800 active:bg-blue-200 focus:ring-blue-300"
-            for="ac-1">Path Usage...</label>
-        <article class="ac-text" id="path-usage-container"></article>
-    </div>
-
-    <hr />
-
-    <div class="flex p-1 shadow-sm">
-        <article id="proxy-check-container" class="text-sm w-full">
-            <div class="ac">
-                <input class="ac-input" id="proxy-status-toggle" name="proxy-status-toggle" type="checkbox" />
-                <label for="proxy-status-toggle" class="ac-label flex items-center" id="proxy-status-label">
-                    <span id="proxy-status-message">Checking proxy status...</span>
-                    <a id="proxy-help-link" href="#" class="text-blue-600 hover:text-blue-800 ml-2 hidden">Need help?</a>
-                </label>
-                <article class="ac-text" id="proxy-details-content">
-                </article>
-            </div>
-        </article>
-    </div>
-
-    <script type="module" src="shared/storage.js"></script>
-    <script type="module" src="popup.js"></script>
-</body>
-
-</html>
