+++ chrome/src/background_helpers/geofence_handler.ts
