+++ chrome/src/background_helpers/request_interception_handler.ts
