+++ chrome/src/background_helpers/tab_handler.ts
