+++ chrome/src/checking/checking.html
@@ -1,0 +1,23 @@
+<!doctype html>
+<html lang="en" class="h-full">
+<head>
+    <link href="../../css/tailwind.min.css" rel="stylesheet">
+    <meta charset="utf-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1"/>
+
+    <title>SCION Check</title>
+</head>
+
+<body class="text-gray-100 m-0 p-8 flex flex-col items-center gap-4 h-full" style="background: #121212;">
+
+    <span id="spinner" class="loading loading-spinner loading-lg"></span>
+
+    <h1 id="title" style="font-size: xx-large">Checking SCION compatibility...</h1>
+
+    <p id="status">Determining the original page you tried to open.</p>
+
+    <p id="original-url" class="font-mono break-all max-w-full text-center"></p>
+
+    <script type="module" src="checking.js"></script>
+</body>
+</html>
