+++ chrome/src/shared/database.ts
@@ -93,14 +93,10 @@ export async function addOrUpdateRequestInDB(entry: RequestSchema): Promise<void
         lastAccessed: now(),
     };
 
-    const addOrUpdateRequest = store.put(newEntry);
-    await requestToPromise(addOrUpdateRequest);
+    store.put(newEntry);
 
     const count = await getEntryCount(store);
-    // since MAX_ENTRIES already accounts for some buffer, concurrent changes to the DNR rules will
-    // with near certainty not exceed the DNR rule limit until rules are evicted, hence simply comparing
-    // count and not e.g. count+10 is sufficient
-    if (count > MAX_ENTRIES) {
+    if (count + 1 > MAX_ENTRIES) {
         await evictLRU(store, EVICT_COUNT);
     }
 
@@ -270,8 +266,11 @@ async function evictLRU(store: IDBObjectStore, count: number): Promise<void> {
  * here.
  */
 async function getEntryCount(store: IDBObjectStore): Promise<number> {
-    const countRequest = store.count();
-    return await requestToPromise(countRequest);
+    const request = store.count();
+    return await new Promise((resolve, reject) => {
+        request.onsuccess = () => resolve(request.result);
+        request.onerror = () => reject(request.error);
+    });
 }
 
 /**
@@ -293,18 +292,5 @@ function transactionDone(transaction: IDBTransaction): Promise<void> {
         transaction.oncomplete = () => resolve();
         transaction.onerror = () => reject(transaction.error);
         transaction.onabort = () => reject(transaction.error);
-    });
-}
-
-/**
- * Converts an {@link IDBRequest} into an await-able {@link Promise}.
- * @example
- * const request = store.put(entry);
- * await requestToPromise(request);
- */
-function requestToPromise<T>(request: IDBRequest<T>): Promise<T> {
-    return new Promise((resolve, reject) => {
-        request.onsuccess = () => resolve(request.result);
-        request.onerror = () => reject(request.error);
     });
 }
