+++ package.json
@@ -6,10 +6,10 @@
   "directories": {
     "doc": "docs"
   },
-"scripts": {
-  "build:css": "tailwindcss -i src/css/tailwind.css -o chrome/css/tailwind.min.css --minify",
-  "watch:css": "tailwindcss -i src/css/tailwind.css -o chrome/css/tailwind.min.css --watch"
-},
+  "scripts": {
+    "build:css": "tailwindcss -i src/css/tailwind.css -o chrome/css/tailwind.min.css --minify",
+    "watch:css": "tailwindcss -i src/css/tailwind.css -o chrome/css/tailwind.min.css --watch"
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/scionproto-contrib/browser-extension.git"
@@ -22,6 +22,7 @@
   },
   "homepage": "https://github.com/scionproto-contrib/browser-extension#readme",
   "devDependencies": {
-    "tailwindcss": "^2.2.19"
+    "daisyui": "^4.12.24",
+    "tailwindcss": "^3.4.0"
   }
 }
